Attempted to Sync Fork + Refactor + Begin Changes
